Use generated zone names for the actual rules
To avoid issues with long chain names when long zone names is
combined with builtin chain names like _POSTROUTING the actual
zone name is no longer used in the chain, but rather a generated
ZONE<ZONE_ID> name.
The Zone ID is just the index of the zone in a list of the zones.
This requires no changes to the configuration, but all valid python
dict keys can now be used as zone name in the configuration,
regardless of underlying iptables chain name limitations.