Make zone to zone rules more efficient and easier to write.
If 'to:' are used instead of FORWARD/INPUT/OUTPUT you can add an
additional level with to zone names and their rules. This makes the
old %{zone} syntax obsolete. Also the ruleset should be more efficient
with multi-interface multi-zone rulesets. The example config
has been updated with an example.

POSTROUTING/PREROUTING is still used in the same way as before even
with to-style syntax.

This change is fully backwards compatible so all current ruleset should
still work.

ADDITIONAL NOTE ABOUT INTRA-ZONE TRAFFIC:
As of this update intra-zone traffic is now ALLOWED by default. This is
more inline of the expected behaviour of a zone firewall.